// types/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
};

/** ACF Link field */
export type AcfLink = {
  __typename?: 'AcfLink';
  /** The target of the link (_blank, etc) */
  target?: Maybe<Scalars['String']>;
  /** The title of the link */
  title?: Maybe<Scalars['String']>;
  /** The url of the link */
  url?: Maybe<Scalars['String']>;
};
// ...
export type Page_Pagepanels_PagePanels_ListList2Col = AcfFieldGroup & {
  __typename?: 'Page_Pagepanels_PagePanels_ListList2Col';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  leftList?: Maybe<Page_Pagepanels_PagePanels_ListList2Col_LeftList>;
  rightList?: Maybe<Page_Pagepanels_PagePanels_ListList2Col_RightList>;
};

/** Field Group */
export type Page_Pagepanels_PagePanels_ListList2Col_LeftList = AcfFieldGroup & {
  __typename?: 'Page_Pagepanels_PagePanels_ListList2Col_LeftList';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  listItems?: Maybe<Array<Maybe<Page_Pagepanels_PagePanels_ListList2Col_LeftList_ListItems>>>;
  listTitle?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Page_Pagepanels_PagePanels_ListList2Col_LeftList_ListItems = AcfFieldGroup & {
  __typename?: 'Page_Pagepanels_PagePanels_ListList2Col_LeftList_listItems';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  listItem?: Maybe<Scalars['String']>;
  listItemLink?: Maybe<AcfLink>;
};

/** Field Group */
export type Page_Pagepanels_PagePanels_ListList2Col_RightList = AcfFieldGroup & {
  __typename?: 'Page_Pagepanels_PagePanels_ListList2Col_RightList';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  listItems?: Maybe<Array<Maybe<Page_Pagepanels_PagePanels_ListList2Col_RightList_ListItems>>>;
  listTitle?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Page_Pagepanels_PagePanels_ListList2Col_RightList_ListItems = AcfFieldGroup & {
  __typename?: 'Page_Pagepanels_PagePanels_ListList2Col_RightList_listItems';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  listItem?: Maybe<Scalars['String']>;
  listItemLink?: Maybe<AcfLink>;
};
// ...
export type Project_Pagepanels_PagePanels_ListList2Col = AcfFieldGroup & {
  __typename?: 'Project_Pagepanels_PagePanels_ListList2Col';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  leftList?: Maybe<Project_Pagepanels_PagePanels_ListList2Col_LeftList>;
  rightList?: Maybe<Project_Pagepanels_PagePanels_ListList2Col_RightList>;
};

/** Field Group */
export type Project_Pagepanels_PagePanels_ListList2Col_LeftList = AcfFieldGroup & {
  __typename?: 'Project_Pagepanels_PagePanels_ListList2Col_LeftList';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  listItems?: Maybe<Array<Maybe<Project_Pagepanels_PagePanels_ListList2Col_LeftList_ListItems>>>;
  listTitle?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Project_Pagepanels_PagePanels_ListList2Col_LeftList_ListItems = AcfFieldGroup & {
  __typename?: 'Project_Pagepanels_PagePanels_ListList2Col_LeftList_listItems';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  listItem?: Maybe<Scalars['String']>;
  listItemLink?: Maybe<AcfLink>;
};

/** Field Group */
export type Project_Pagepanels_PagePanels_ListList2Col_RightList = AcfFieldGroup & {
  __typename?: 'Project_Pagepanels_PagePanels_ListList2Col_RightList';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  listItems?: Maybe<Array<Maybe<Project_Pagepanels_PagePanels_ListList2Col_RightList_ListItems>>>;
  listTitle?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Project_Pagepanels_PagePanels_ListList2Col_RightList_ListItems = AcfFieldGroup & {
  __typename?: 'Project_Pagepanels_PagePanels_ListList2Col_RightList_listItems';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  listItem?: Maybe<Scalars['String']>;
  listItemLink?: Maybe<AcfLink>;
};